# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RTCAS API",
    "version": "1.0.0",
    "description": "RTCAS Backend API Definition"
  },
  "servers": [
    { "url": "http://localhost:3000" },
    { "url": "https://api.rtcas.in.th" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/knowledge-assessments/import": {
      "post": {
        "summary": "Import knowledge assessments from Excel",
        "tags": ["Knowledge Assessments"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Imported" },
          "400": { "description": "Invalid file or data" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/knowledge-assessments": {
      "get": {
        "summary": "List knowledge assessments",
        "tags": ["Knowledge Assessments"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": { "type": "integer" }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "List returned" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      },
      "delete": {
        "summary": "Delete all knowledge assessments",
        "tags": ["Knowledge Assessments"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Deleted" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/knowledge-assessments/{id}": {
      "delete": {
        "summary": "Delete knowledge assessment by id",
        "tags": ["Knowledge Assessments"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found" }
        }
      }
    },
...
    "/api/admin/news": {
      "post": {
        "summary": "Create news (admin)",
        "tags": ["News"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      },
      "get": {
        "summary": "List all news (admin)",
        "tags": ["News"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List returned" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/admin/news/{id}": {
      "put": {
        "summary": "Update news (admin)",
        "tags": ["News"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": { "type": "string", "format": "binary" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated" },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "summary": "Delete news (admin)",
        "tags": ["News"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found" }
        }
      }
    },
...
    "/api/news/{id}": {
      "get": {
        "summary": "Get news by id (public)",
        "tags": ["News"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "News returned" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/owner/teacher-leaves": {
      "get": {
        "summary": "List teacher leave requests (owner)",
        "tags": ["Owner"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List returned" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/owner/teacher-leaves/{id}/status": {
      "patch": {
        "summary": "Update teacher leave status (owner)",
        "tags": ["Owner"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": { "description": "Updated" },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/owner/official-duty-leaves": {
      "get": {
        "summary": "List official duty leave requests (owner)",
        "tags": ["Owner"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "List returned" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/owner/official-duty-leaves/{id}/status": {
      "patch": {
        "summary": "Update official duty leave status (owner)",
        "tags": ["Owner"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "200": { "description": "Updated" },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/api/owner/notifications": {
      "get": {
        "summary": "Get owner notifications",
        "tags": ["Owner Notifications"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": { "type": "integer" }
          },
          {
            "in": "query",
            "name": "pageSize",
            "required": false,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Notifications returned" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
    "/api/owner/notifications/read": {
      "patch": {
        "summary": "Mark owner notifications as read",
        "tags": ["Owner Notifications"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ids"],
                "properties": {
                  "ids": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Marked as read" },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" }
        }
      }
    },
...
  },
...
}